requirements.txt
absl-py==0.7.1
astor==0.7.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
catkin-pkg==0.4.10
certifi==2018.11.29
cffi==1.13.2
chardet==3.0.4
cycler==0.10.0
Cython==0.29.6
DateTime==4.3
decorator==4.3.0
defusedxml==0.5.0
e==1.4.5
entrypoints==0.3
envs==1.3
future==0.17.1
gast==0.2.2
gitdb2==2.0.6
GitPython==3.0.4
glfw==1.9.1
grpcio==1.20.0
gym==0.10.9
h5py==2.9.0
hjson==3.0.1
idna==2.8
imageio==2.6.1
importlib-metadata==1.5.0
ipdb==0.12.2
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lockfile==0.12.2
lxml==4.4.2
Markdown==3.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mock==2.0.0
more-itertools==8.2.0
mpmath==1.1.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.0
packaging==20.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.1
pathlib2==2.3.5
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.13.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyglet==1.3.2
Pygments==2.3.1
pyparsing==2.3.1
pyquaternion==0.9.5
pytest==5.3.5
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
smmap2==2.0.5
snakeviz==2.0.1
sympy==1.3
tensorboard==1.12.2
tensorboardX==1.6
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
torch==1.1.0
torchsummary==1.5.1
torchvision==0.3.0
tornado==5.1.1
tqdm==4.36.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
